+++ R/ShinyGUIClass.R
@@ -8,6 +8,10 @@ ShinyGUI <- setRefClass("ShinyGUI",
                                     demographics = 'list',
                                     lastpage = 'function',
                                     instructions = 'character',
+                                    itemtimer = 'character',
+                                    incorrect = 'character',
+                                    failpass = 'character',
+                                    timemsg = 'character',
                                     begin_message = 'character',
                                     stem_locations = 'character',
                                     stem_expressions = 'character',
@@ -38,7 +42,7 @@ ShinyGUI <- setRefClass("ShinyGUI",
                               forced_choice <<- TRUE
                               theme <<- ''
                               Timer <- as.numeric(Timer)
-                              timer <<- ifelse(is.finite(Timer), Timer, as.numeric(NA))
+                              timer <<- ifelse(is.finite(Timer), Timer, 0)
                               if(length(CustomTypes)){
                                   if(length(CustomTypes) != length(unique(names(CustomTypes))))
                                       stop('customTypes list requires unique names for each function', call.=FALSE)
@@ -76,7 +80,11 @@ ShinyGUI <- setRefClass("ShinyGUI",
                               author <<- 'Author information'
                               instructions <<- c("To progress through the interface, click on the action button below.",
                                                  "Next")
-                              time_remaining <<- "Time remaining: "
+                              itemtimer <<- 'Item timer: '
+                              incorrect <<- 'The answer provided was incorrect. Please select an alternative.'
+                              failpass <<- 'Incorrect Login Name/Password. Please try again (you have %s attempts remaining).'
+                              timemsg <<- c('hour ','minutes ','seconds ', 'and ')
+                              time_remaining <<- 'Time remaining: '
                               response_msg <<- "Please provide a suitable response"
                               demographic_inputIDs <<- character(0)
                               max_password_attempts <<- 3L
@@ -97,7 +105,7 @@ ShinyGUI <- setRefClass("ShinyGUI",
                                                  
                               if(length(shinyGUI)){
                                   dnames <- names(shinyGUI)
-                                  gnames <- c('title', 'authors', 'instructions', 'firstpage', 'demographics',
+                                  gnames <- c('title', 'authors', 'instructions', 'itemtimer', 'incorrect','failpass','timemsg', 'firstpage', 'demographics',
                                               'demographics_inputIDs', 'temp_file', "time_remaining", "response_msg",
                                               'lastpage', 'css', 'stem_dims', 'forced_choice', 'stem_locations',
                                               'begin_message', 'ui', 'password', 'stem_default_format',
@@ -112,6 +120,14 @@ ShinyGUI <- setRefClass("ShinyGUI",
                                       theme <<- shinyGUI$theme
                                   if(!is.null(shinyGUI$instructions))
                                       instructions <<- shinyGUI$instructions
+                                  if(!is.null(shinyGUI$itemtimer))
+                                      itemtimer <<- shinyGUI$itemtimer
+                                  if(!is.null(shinyGUI$incorrect))
+                                      incorrect <<- shinyGUI$incorrect
+                                  if(!is.null(shinyGUI$failpass))
+                                      failpass <<- shinyGUI$failpass
+                                  if(!is.null(shinyGUI$timemsg))
+                                      timemsg <<- shinyGUI$timemsg
                                   if(!is.null(shinyGUI$begin_message))
                                       begin_message <<- shinyGUI$begin_message
                                   if(!is.null(shinyGUI$title))
@@ -143,9 +159,8 @@ ShinyGUI <- setRefClass("ShinyGUI",
                                   if(!is.null(shinyGUI$max_password_attempts))
                                       max_password_attempts <<- shinyGUI$max_password_attempts
                               }
-                              if(any(!is.na(timer)) && forced_choice) 
-                                  stop('Timer inputs cannot be combined with shinyGUI$forced_choice = TRUE', 
-                                       call.=FALSE)
+                              if(any(timer > 0)) 
+                                  forced_choice <<- FALSE
                           })
                       
 )